Move relay hints into relay module
