+++ web/src/lib/nostr/relay/relay-hints.ts
